repo Корректировка моделей, маппинга, бизнес логики

+++ Application/BussinessLogic/Hotel/HotelBussinessLogic.cs
@@ -1,57 +1,22 @@
 ﻿using Application.BussinessLogic.GeneralMethods;
 using Application.DTO.Hotel;
-using Application.DTO.Room;
 using Application.Interfaces.Repository;
 using AutoMapper;
-using AutoMapper.QueryableExtensions;
-using Domain.Models;
-using Microsoft.EntityFrameworkCore;
-using FoodEntity = Domain.Models.Food;
 using HotelEntity = Domain.Models.Hotel;
-using HotelFacilityEntity = Domain.Models.HotelFacility;
-using HotelPhotoEntity = Domain.Models.HotelPhoto;
-using LocationEntity = Domain.Models.Location;
-using PriceEntity = Domain.Models.Price;
-using ReviewEntity = Domain.Models.Review;
-using RoomEntity = Domain.Models.Room;
 
 namespace Application.BussinessLogic.Hotel
 {
   public class HotelBussinessLogic : IHotelBussinessLogic
   {
-    //private readonly IRepositoryBase<HotelEntity> _repositoryHotel;
-    //private readonly IRepositoryBase<RoomEntity> _repositoryRoom;
-    //private readonly IRepositoryBase<FoodEntity> _repositoryFood;
-    //private readonly IRepositoryBase<HotelFacilityEntity> _repositoryHotelFacility;
-    //private readonly IRepositoryBase<HotelPhotoEntity> _repositoryHotelPhoto;
-    //private readonly IRepositoryBase<LocationEntity> _repositoryLocation;
-    //private readonly IRepositoryBase<PriceEntity> _repositoryPrice;
-    //private readonly IRepositoryBase<ReviewEntity> _repositoryReview;
     private readonly IGeneralBussinessLogic _generalBussinessLogic;
     private readonly IRepositoryManager _repositoryManager;
 
     private readonly IMapper _mapper;
 
-    public HotelBussinessLogic(/*IRepositoryBase<HotelEntity> repositoryHotel,
-                               IRepositoryBase<RoomEntity> repositoryRoom,
-                               IRepositoryBase<FoodEntity> repositoryFood,
-                               IRepositoryBase<HotelFacilityEntity> repositoryHotelFacility,
-                               IRepositoryBase<HotelPhotoEntity> repositoryHotelPhoto,
-                               IRepositoryBase<LocationEntity> repositoryLocation,
-                               IRepositoryBase<PriceEntity> repositoryPrice,
-                               IRepositoryBase<ReviewEntity> repositoryReview,*/
-                               IGeneralBussinessLogic generalBussinessLogic,
+    public HotelBussinessLogic(IGeneralBussinessLogic generalBussinessLogic,
                                IMapper mapper,
                                IRepositoryManager repositoryManager)
     {
-      //_repositoryHotel = repositoryHotel;
-      //_repositoryRoom = repositoryRoom;
-      //_repositoryFood = repositoryFood;
-      //_repositoryHotelFacility = repositoryHotelFacility;
-      //_repositoryHotelPhoto = repositoryHotelPhoto;
-      //_repositoryLocation = repositoryLocation;
-      //_repositoryPrice = repositoryPrice;
-      //_repositoryReview = repositoryReview;
       _generalBussinessLogic = generalBussinessLogic;
       _mapper = mapper;
       _repositoryManager = repositoryManager;
@@ -59,18 +24,8 @@ namespace Application.BussinessLogic.Hotel
 
     public async Task<List<HotelAllDto>> GetAllAsync()
     {
-      //return await _repositoryHotel.FindAll()
-      //                             .ProjectTo<HotelAllDto>(_mapper.ConfigurationProvider)
-      //                             .ToListAsync();
-
-      //var listDB = await _repositoryManager.HotelRepository.GetAllAsync(trackChanges: true);
-      //var listReturn = listDB.AsQueryable().ProjectTo<HotelAllDto>(_mapper.ConfigurationProvider);
-
-      //return null;
-
-      return await _repositoryManager.HotelRepository.GetAll(trackChanges: true).AsQueryable()
-                                  .ProjectTo<HotelAllDto>(_mapper.ConfigurationProvider)
-                                  .ToListAsync();
+      var listDB = _mapper.Map<List<HotelAllDto>>(await _repositoryManager.HotelRepository.GetAllAsync(trackChanges: true));
+      return listDB.ToList();
     }
 
     public async Task<Guid> CreateAsync(HotelCreateDto dto)
@@ -91,46 +46,36 @@ namespace Application.BussinessLogic.Hotel
         price.HotelId = entity.Id;
       //await _repositoryHotel.CreateAsync(entity);
       await _repositoryManager.HotelRepository.CreateEntityAsync(entity);
+      await _repositoryManager.SaveAsync();
       return entity.Id;
     }
 
     public async Task<HotelDto> GetByIdAsync(Guid id)
     {
-      //var entityDto = await _repositoryHotel.FindAll()
-      //                     .ProjectTo<HotelDto>(_mapper.ConfigurationProvider)
-      //                     .FirstOrDefaultAsync(x => x.Id == id);
-
-      var entityDto = await _repositoryManager.HotelRepository.GetByCondition(x => x.Id == id, trackChanges: true)
-                           .AsQueryable().ProjectTo<HotelDto>(_mapper.ConfigurationProvider).FirstAsync();
-      return entityDto;
+      var hotel = await _repositoryManager.HotelRepository.GetOneAsync(x => x.Id == id, trackChanges: true);
+      if (hotel == null)
+        return new HotelDto();
+      else
+        return  _mapper.Map<HotelDto>(hotel);
     }
 
     public async Task DeleteAsync(Guid hotelId)
     {
-      //var hotel = await _repositoryHotel.FindOneAsync(x => x.Id == hotelId);
-      var hotel = await _repositoryManager.HotelRepository.GetOneAsync(x => x.Id == hotelId);
+      var hotel = await _repositoryManager.HotelRepository.GetOneAsync(x => x.Id == hotelId, trackChanges: true);
 
-      //await _repositoryRoom.DeleteRangeAsync(hotel.Rooms);
-      //await _repositoryFood.DeleteRangeAsync(hotel.Foods);
-      //await _repositoryHotelFacility.DeleteRangeAsync(hotel.HotelFacilities);
-      //await _repositoryHotelPhoto.DeleteRangeAsync(hotel.HotelPhotos);
-      //await _repositoryLocation.DeleteRangeAsync(hotel.Locations);
-      //await _repositoryPrice.DeleteRangeAsync(hotel.Prices);
-      //await _repositoryReview.DeleteRangeAsync(hotel.Reviews);
+      //await _repositoryManager.RoomRepository.DeleteEntityRangeAsync(hotel.Rooms);
+      //await _repositoryManager.FoodRepository.DeleteEntityRangeAsync(hotel.Foods);
+      //await _repositoryManager.HotelFacilityRepository.DeleteEntityRangeAsync(hotel.HotelFacilities);
+      //await _repositoryManager.HotelPhotoRepository.DeleteEntityRangeAsync(hotel.HotelPhotos);
+      //await _repositoryManager.LocationRepository.DeleteEntityRangeAsync(hotel.Locations);
+      //await _repositoryManager.PriceRepository.DeleteEntityRangeAsync(hotel.Prices);
+      //await _repositoryManager.ReviewRepository.DeleteEntityRangeAsync(hotel.Reviews);
 
-      await _repositoryManager.RoomRepository.DeleteEntityRangeAsync(hotel.Rooms);
-      await _repositoryManager.FoodRepository.DeleteEntityRangeAsync(hotel.Foods);
-      await _repositoryManager.HotelFacilityRepository.DeleteEntityRangeAsync(hotel.HotelFacilities);
-      await _repositoryManager.HotelPhotoRepository.DeleteEntityRangeAsync(hotel.HotelPhotos);
-      await _repositoryManager.LocationRepository.DeleteEntityRangeAsync(hotel.Locations);
-      await _repositoryManager.PriceRepository.DeleteEntityRangeAsync(hotel.Prices);
-      await _repositoryManager.ReviewRepository.DeleteEntityRangeAsync(hotel.Reviews);
-
-      //_repositoryHotel.Delete(hotel);
-      //await _repositoryRoom.SaveAsync();
-
-      _repositoryManager.HotelRepository.DeleteEntity(hotel);
-      await _repositoryManager.SaveAsync();
+      if (hotel is not null)
+      {
+        _repositoryManager.HotelRepository.DeleteEntity(hotel);
+        await _repositoryManager.SaveAsync();
+      }
     }
 
     public async Task UpdateAsync(HotelUpdateDto dto)
@@ -140,7 +85,7 @@ namespace Application.BussinessLogic.Hotel
       if (entity == null) return;
 
       _mapper.Map(dto, entity);
-
+      
       await _generalBussinessLogic.UpdateCollectionAsync(
             entity.Prices,
             dto.Prices,
@@ -200,7 +145,7 @@ namespace Application.BussinessLogic.Hotel
             location => location.Id,
             dto => dto.Id
             );
-
+      
       //await _repositoryHotel.SaveAsync();
       await _repositoryManager.SaveAsync();
     }
